Move parseUri to thunk
