+++ src/modules/UI/scenes/Scan/Scan.ui.js
@@ -1,9 +1,9 @@
 // @flow
 
-import type { EdgeCurrencyWallet, EdgeParsedUri, EdgeTokenInfo } from 'edge-core-js'
 import React, { Component } from 'react'
-import { ActivityIndicator, Alert, Text, TouchableHighlight, View } from 'react-native'
+import { ActivityIndicator, Text, TouchableHighlight, View } from 'react-native'
 import Camera from 'react-native-camera'
+
 // $FlowFixMe
 import ImagePicker from 'react-native-image-picker'
 import { Actions } from 'react-native-router-flux'
@@ -12,11 +12,8 @@ import Ionicon from 'react-native-vector-icons/Ionicons'
 
 import * as Constants from '../../../../constants/indexConstants'
 import s from '../../../../locales/strings.js'
-import type { GuiWallet } from '../../../../types'
-import * as WALLET_API from '../../../Core/Wallets/api.js'
 import type { PermissionStatus } from '../../../ReduxTypes'
 import WalletListModal from '../../../UI/components/WalletListModal/WalletListModalConnector'
-import * as UTILS from '../../../utils.js'
 import ABAlert from '../../components/ABAlert/indexABAlert'
 import T from '../../components/FormattedText'
 import Gradient from '../../components/Gradient/Gradient.ui'
@@ -27,25 +24,12 @@ import styles, { styles as styleRaw } from './style'
 
 type Props = {
   cameraPermission: PermissionStatus,
-  edgeWallet: EdgeCurrencyWallet,
   torchEnabled: boolean,
   scanEnabled: boolean,
-  walletListModalVisible: boolean,
-  scanToWalletListModalVisibility: boolean,
-  scanFromWalletListModalVisibility: boolean,
-  scanToWalletListModalVisibility: boolean,
-  walletId: string,
-  guiWallet: GuiWallet,
-  dispatchEnableScan(): void,
-  dispatchDisableScan(): void,
   toggleEnableTorch(): void,
   toggleAddressModal(): void,
-  toggleWalletListModal(): void,
   toggleScanToWalletListModal(): void,
-  updateParsedURI(EdgeParsedUri): void,
-  loginWithEdge(string): void,
-  toggleScanToWalletListModal: () => void,
-  routeToAddToken: EdgeTokenInfo => void
+  parseUri: (data: string) => void
 }
 
 const HEADER_TEXT = s.strings.send_scan_header_text
@@ -122,48 +106,11 @@ export default class Scan extends Component<Props> {
   onBarCodeRead = (scan: { data: string }) => {
     if (!this.props.scanEnabled) return
     const uri = scan.data
-    this.parseURI(uri)
+    this.parseUri(uri)
   }
 
-  parseURI = (uri: string) => {
-    try {
-      if (/^airbitz:\/\/edge\//.test(uri)) {
-        this.props.loginWithEdge(uri)
-        return
-      }
-      const parsedURI = WALLET_API.parseURI(this.props.edgeWallet, uri)
-      if (parsedURI.token) {
-        // token URI, not pay
-        const { contractAddress, currencyName, multiplier } = parsedURI.token
-        const currencyCode = parsedURI.token.currencyCode.toUpperCase()
-        const wallet = this.props.guiWallet
-        const walletId = this.props.guiWallet.id
-        let decimalPlaces = 18
-        if (parsedURI.token && parsedURI.token.multiplier) {
-          decimalPlaces = UTILS.denominationToDecimalPlaces(parsedURI.token.multiplier)
-        }
-        const parameters = {
-          contractAddress,
-          currencyCode,
-          currencyName,
-          multiplier,
-          decimalPlaces,
-          walletId,
-          wallet,
-          onAddToken: UTILS.noOp
-        }
-        Actions.addToken(parameters)
-      } else {
-        // assume pay URI
-        this.props.updateParsedURI(parsedURI)
-        Actions.sendConfirmation('fromScan')
-      }
-    } catch (error) {
-      this.props.dispatchDisableScan()
-      Alert.alert(s.strings.fragment_send_send_bitcoin_unscannable, error.toString(), [
-        { text: s.strings.string_ok, onPress: () => this.props.dispatchEnableScan() }
-      ])
-    }
+  parseUri = (data: string) => {
+    this.props.parseUri(data)
   }
 
   selectPhotoTapped = () => {
